+++ docs/changelog.rst
@@ -6,6 +6,9 @@ Changelog
 `Unreleased`_
 -------------
 
+`0.9.0`_ - 2020-08-13
+---------------------
+
 Added
 ~~~~~
 
@@ -142,7 +145,8 @@ Added
 
 - Initial public release
 
-.. _Unreleased: https://github.com/kiwicom/pytest-recording/compare/v0.8.1...HEAD
+.. _Unreleased: https://github.com/kiwicom/pytest-recording/compare/v0.9.0...HEAD
+.. _0.9.0: https://github.com/kiwicom/pytest-recording/compare/v0.8.1...v0.9.0
 .. _0.8.1: https://github.com/kiwicom/pytest-recording/compare/v0.8.0...v0.8.1
 .. _0.8.0: https://github.com/kiwicom/pytest-recording/compare/v0.7.0...v0.8.0
 .. _0.7.0: https://github.com/kiwicom/pytest-recording/compare/v0.6.0...v0.7.0